extract/refactor setup [ init , render ]
