+++ sun/index.html
@@ -13,7 +13,6 @@
     </style>
   </head>
   <body></body>
-  <script src="../../suncalc.js"></script>
-
+  
   <script type="module" src="./js/index.js"></script>
 </html>
