+++ suncalc.js
@@ -4,7 +4,7 @@
  https://github.com/mourner/suncalc
 */
 
-(function () { 'use strict';
+ 'use strict';
 
 // shortcuts for easier to read formulas
 
@@ -166,8 +166,8 @@ SunCalc.getTimes = function (date, lat, lng) {
 };
 
 // export as Node module / AMD module / browser variable
-if (typeof exports === 'object' && typeof module !== 'undefined') module.exports = SunCalc;
-else if (typeof define === 'function' && define.amd) define(SunCalc);
-else window.SunCalc = SunCalc;
+// if (typeof exports === 'object' && typeof module !== 'undefined') export {SunCalc};
+// else if (typeof define === 'function' && define.amd) define(SunCalc);
+// else window.SunCalc = SunCalc;
+export {SunCalc}
 
-}());
